Value::Path type plumbing (phase 2)
Adds the `Path` value variant, widens the type-flag space to u32 with a 17-wide operator
dispatch table, `ArgType::Path` (`|p: path|` hints, "path" in signatures), calling a path as
`p(t)` / `p(t=)`, and Path arms for Debug, JSON, retained-size, and the optimizer's content
hashing / purity checks. No builtin uses the type yet.

+++ src/viz/wasm/geoscript/src/builtins/fn_defs.rs
@@ -27,7 +27,7 @@ pub struct ArgDef {
   /// This will be populated lazily
   pub interned_name: Sym,
   /// Bitflags
-  pub valid_types: u16,
+  pub valid_types: u32,
   pub default_value: DefaultValue,
   pub description: &'static str,
 }
